fix(load): verify ptr map helpers

+++ game/LOAD/LOAD_16_DramFileCallback.c
@@ -20,7 +20,7 @@ void LOAD_DramFileCallback(struct LoadQueueSlot *lqs)
 	{
 		struct DramPointerMap *dpm = (struct DramPointerMap *)&realFileBuf[ptrMapOffset];
 
-		LOAD_RunPtrMap((int)realFileBuf, (int *)DRAM_GETOFFSETS(dpm), dpm->numBytes >> 2);
+		LOAD_RunPtrMap(realFileBuf, (int *)DRAM_GETOFFSETS(dpm), dpm->numBytes >> 2);
 
 		DRAM_SET_PATCHED(fileBuf);
